pre-primary and primary education variables
