+++ etl/steps/data/garden/wb/2023-07-10/education_primary_pre_primary.countries.json
@@ -217,5 +217,6 @@
   "Yemen, Rep.": "Yemen",
   "Zambia": "Zambia",
   "Zimbabwe": "Zimbabwe",
-  "East Asia & Pacific": "East Asia & Pacific"
+  "East Asia & Pacific": "East Asia & Pacific",
+  "Korea, Dem. People’s Rep.": "North Korea"
 }
