+++ etl/steps/data/grapher/wb/2023-07-10/education_primary_pre_primary.py
@@ -1,0 +1,39 @@
+"""Load a garden dataset and create a grapher dataset."""
+
+from typing import cast
+
+from owid.catalog import Dataset
+
+from etl.helpers import PathFinder, create_dataset, grapher_checks
+
+# Get paths and naming conventions for current step.
+paths = PathFinder(__file__)
+
+
+def run(dest_dir: str) -> None:
+    #
+    # Load inputs.
+    #
+    # Load garden dataset.
+    ds_garden = cast(Dataset, paths.load_dependency("education_primary_pre_primary"))
+
+    # Read table from garden dataset.
+    tb = ds_garden["education_primary_pre_primary"]
+
+    #
+    # Process data.
+    #
+
+    #
+    # Save outputs.
+    #
+    # Create a new grapher dataset with the same metadata as the garden dataset.
+    ds_grapher = create_dataset(dest_dir, tables=[tb], default_metadata=ds_garden.metadata)
+
+    #
+    # Checks.
+    #
+    grapher_checks(ds_grapher)
+
+    # Save changes in the new grapher dataset.
+    ds_grapher.save()
